[agent] refactor(assembly): extract error + recognizers modules from lib.rs
