+++ crates/gaze-assembly/src/error.rs
@@ -1,0 +1,11 @@
+use thiserror::Error;
+
+#[derive(Debug, Error)]
+pub enum BuildError {
+    #[error("policy error: {0}")]
+    Policy(#[from] gaze::PolicyError),
+    #[error("rulepack error: {0}")]
+    Rulepack(#[from] gaze::RulepackError),
+    #[error("pipeline error: {0}")]
+    Pipeline(#[from] gaze::Error),
+}
